Adds ET parfile and ccl files
Please note that the current implementation is very ugly due to the
void pointers from Cactus. In the future, the interface.ccl will
*not* be like this. For now, however, I am doing the simplest
implementation with the wrapper functions in con2prim_ETK_interface.c
which cast these to the proper types and pass them to the normal functions.
These allow me to easily call them from IllinoisGRMHD for testing.