Added many comments, simplified function arguments
Many functions, particularly initialize functions, had if
statements to only set variables like temperature if the EOS was
tabulated. However, the argument list has temperature regardless,
which will be junk for a Hybrid simulation. Thus, we can just set
it to junk, as it will/should not be used, thereby removing the
need to pass the eos or params structs.

In addition, I also split many functions into individual files and
improved the directory scheme in EOS.